+++ libs/ghostos/ghostos/demo/moss_libs/self_updater_moss.py
@@ -1,0 +1,6 @@
+from ghostos_moss.abcd import SelfUpdater, Moss as Parent
+
+
+class Moss(Parent):
+    updater: SelfUpdater
+    """update current module code"""