+++ libs/moss/src/ghostos_moss/examples/random_moss.py
@@ -1,0 +1,5 @@
+from ghostos_moss.abcd import Moss as Parent
+
+
+class RandomMoss(Parent):
+    pass