+++ libs/moss/src/ghostos_moss/examples/self_update_moss.py
@@ -1,0 +1,5 @@
+from ghostos_moss.abcd import Moss as Parent, SelfUpdater
+
+
+class Moss(Parent):
+    updater: SelfUpdater