+++ libs/common/RELEASES.md
@@ -1,8 +1,14 @@
+# 0.3.0 
+
+move `ghostos.prompter` to `ghostos_common.prompter`.
+provide prompt object model for `ghosots_moss`.
+
+
 # 0.2.0
 
-# 0.2.3
+## 0.2.3
 
 fix the tree-sitter parser bugs.
 
-# 0.2.1
+## 0.2.1
 prepare uv instead poetry.